[api][webui] extract authentication code to Authenticator class
