+++ src/api/app/controllers/application_controller.rb
@@ -5,28 +5,17 @@ require_dependency 'opensuse/permission'
 require_dependency 'opensuse/backend'
 require_dependency 'opensuse/validator'
 require_dependency 'api_exception'
+require_dependency 'authenticator'
 
 class ApplicationController < ActionController::Base
   include Pundit
   protect_from_forgery
-
-  include ForbidsAnonymousUsers
-
-  class NoDataEntered < APIException
-    setup 403
-  end
-
-  class AuthenticationRequiredError < APIException
-    setup 401, "Authentication required"
-  end
 
   include ActionController::ImplicitRender
   include ActionController::MimeResponds
 
   # session :disabled => true
 
-  @user_permissions = nil
-  @http_user = nil
   @skip_validation = false
 
   before_action :validate_xml_request, :add_api_version
@@ -41,11 +30,29 @@ class ApplicationController < ActionController::Base
   before_action :validate_params
   before_action :require_login
 
-  # contains current authentification method, one of (:proxy, :basic)
-  attr_accessor :auth_method
+  delegate :extract_user,
+           :extract_user_public,
+           :require_login,
+           :require_admin,
+           :auth_method,
+           to: :authenticator
+
+  def authenticator
+    @authenticator ||= Authenticator.new(request, session)
+  end
 
   def pundit_user
     return User.current unless User.current.is_nobody?
+  end
+
+  def permissions
+    authenticator.user_permissions
+  end
+
+  # TODO: There are currently two ways of accessing the logged in user: User.curent and user
+  #       We should pick only one of them to use.
+  def user
+    authenticator.http_user
   end
 
   # Method for mapping actions in a controller to (XML) schemas based on request
@@ -78,21 +85,6 @@ class ApplicationController < ActionController::Base
 
   protected
 
-  def load_nobody
-    @http_user = User.find_nobody!
-    User.current = @http_user
-    @user_permissions = Suse::Permission.new( User.current )
-  end
-
-  def require_admin
-    logger.debug "Checking for  Admin role for user #{@http_user.login}"
-    unless @http_user.is_admin?
-      logger.debug "not granted!"
-      render_error(status: 403, errorcode: "put_request_no_permission", message: "Requires admin privileges") && (return false)
-    end
-    true
-  end
-
   def validate_params
     params.each do |key, value|
       next if value.nil?
@@ -102,147 +94,6 @@ class ApplicationController < ActionController::Base
       end
     end
     true
-  end
-
-  class InactiveUserError < APIException
-    setup 403
-  end
-
-  class UnconfirmedUserError < APIException
-    setup 403
-  end
-
-  class UnregisteredUserError < APIException
-    setup 403
-  end
-
-  def extract_proxy_user
-    @auth_method = :proxy
-    proxy_user = request.env['HTTP_X_USERNAME']
-    if proxy_user
-      logger.info "iChain user extracted from header: #{proxy_user}"
-    end
-
-    # we're using a login proxy, there is no need to authenticate the user from the credentials
-    # However we have to care for the status of the user that must not be unconfirmed or proxy requested
-    if proxy_user
-      @http_user = User.find_by_login proxy_user
-
-      # If we do not find a User here, we need to create a user and wait for
-      # the confirmation by the user and the BS Admin Team.
-      unless @http_user
-        if ::Configuration.registration == "deny"
-          logger.debug("No user found in database, creation disabled")
-          raise AuthenticationRequiredError.new "User '#{login}' does not exist"
-        end
-        # Generate and store a fake pw in the OBS DB that no-one knows
-        # FIXME: we should allow NULL passwords in DB, but that needs user management cleanup
-        chars = ["A".."Z", "a".."z", "0".."9"].collect(&:to_a).join
-        fakepw = (1..24).collect { chars[rand(chars.size)] }.pack("a" * 24)
-        @http_user = User.new(
-            login: proxy_user,
-            state: User.default_user_state,
-            password: fakepw)
-      end
-
-      # update user data from login proxy headers
-      @http_user.update_user_info_from_proxy_env(request.env) if @http_user
-    else
-      logger.error "No X-username header from login proxy! Are we really using an authentification proxy?"
-    end
-  end
-
-  def authorization_infos
-    # 1. try to get it where mod_rewrite might have put it
-    # 2. for Apace/mod_fastcgi with -pass-header Authorization
-    # 3. regular location
-    %w{X-HTTP_AUTHORIZATION Authorization HTTP_AUTHORIZATION}.each do |header|
-      if request.env.has_key? header
-        return request.env[header].to_s.split
-      end
-    end
-    return
-  end
-
-  def extract_basic_auth_user
-    authorization = authorization_infos
-
-    # privacy! logger.debug( "AUTH: #{authorization.inspect}" )
-
-    if authorization && authorization[0] == "Basic"
-      # logger.debug( "AUTH2: #{authorization}" )
-      @login, @passwd = Base64.decode64(authorization[1]).split(':', 2)[0..1]
-
-      # set password to the empty string in case no password is transmitted in the auth string
-      @passwd ||= ""
-    else
-      logger.debug "no authentication string was sent"
-    end
-  end
-
-  def extract_user
-    mode = CONFIG['proxy_auth_mode'] || CONFIG['ichain_mode'] || :basic
-    if mode == :on
-      extract_proxy_user
-    else
-      @auth_method = :basic
-
-      extract_basic_auth_user
-
-      @http_user = User.find_with_credentials @login, @passwd if @login
-    end
-
-    if !@http_user && session[:login]
-      @http_user = User.find_by_login session[:login]
-    end
-
-    check_extracted_user
-  end
-
-  def check_for_anonymous_user
-    if ::Configuration.anonymous
-      # Fixed list of clients which do support the read only mode
-      hua = request.env['HTTP_USER_AGENT']
-      if hua # ignore our test suite (TODO: we need to fix that)
-        load_nobody
-        return true
-      end
-    end
-    false
-  end
-
-  def require_login
-    # we allow anonymous user only for rare special operations (if configured) but we require
-    # a valid account for all other operations.
-    # For this rare special operations we simply skip the require login before filter!
-    # At the moment these operations are the /public, /trigger and /about controller actions.
-    be_not_nobody!
-  end
-
-  def check_extracted_user
-    unless @http_user
-      if @login.blank?
-        return true if check_for_anonymous_user
-        raise AuthenticationRequiredError.new
-      end
-      raise AuthenticationRequiredError.new "Unknown user '#{@login}' or invalid password"
-    end
-
-    if @http_user.state == 'unconfirmed'
-      raise UnconfirmedUserError.new "User is registered but not yet approved. " +
-                                         "Your account is a registered account, but it is not yet approved for the OBS by admin."
-    end
-
-    User.current = @http_user
-
-    if @http_user.state == 'confirmed'
-      logger.debug "USER found: #{@http_user.login}"
-      @user_permissions = Suse::Permission.new(@http_user)
-      return true
-    end
-
-    raise InactiveUserError.new "User is registered but not in confirmed state. Your account is a registered account, " +
-                                "but it is in a not active state."
   end
 
   def require_valid_project_name
@@ -428,14 +279,6 @@ class ApplicationController < ActionController::Base
                  message: message
   end
 
-  def permissions
-    @user_permissions
-  end
-
-  def user
-    @http_user
-  end
-
   def require_parameter!(parameter)
     raise MissingParameterError, "Required Parameter #{parameter} missing" unless params.include? parameter.to_s
   end
@@ -584,15 +427,6 @@ class ApplicationController < ActionController::Base
 
   def set_response_format_to_xml
     request.format = :xml if request.format == :html
-  end
-
-  def extract_user_public
-    if ::Configuration.anonymous
-      load_nobody # to become _public_ special user
-    else
-      logger.error 'No public access is configured'
-      render_error( message: 'No public access is configured', status: 401 )
-    end
   end
 
   private
