--- src/api/app/mixins/forbids_anonymous_users.rb
@@ -1,9 +1,0 @@
-module ForbidsAnonymousUsers
-  class AnonymousUser < APIException
-    setup 401
-  end
-
-  def be_not_nobody!
-    raise AnonymousUser.new 'Anonymous user is not allowed here - please login' if !User.current || User.current.is_nobody?
-  end
-end
